final refinement
