+++ apps/client/.env.sample
@@ -1,0 +1,1 @@
+VITE_BACKEND_URL="http://localhost:4000"
