+++ apps/client/src/api/trpc.ts
@@ -1,3 +1,4 @@
+import Configs from '@/lib/configs';
 import type {AppRouter} from '../../../server/src';
 import {createTRPCReact, httpBatchLink} from '@trpc/react-query';
 
@@ -7,7 +8,7 @@ export const trpc = createTRPCReact<AppRouter>();
 export const trpcClient = trpc.createClient({
     links:[
         httpBatchLink({
-            url:"http://localhost:4000/trpc"
+            url:`${Configs.BACKEND_URL}/trpc`
         })
     ]
 })
