+++ apps/client/src/lib/configs.ts
@@ -1,0 +1,5 @@
+ class Configs{
+    public static BACKEND_URL:string = import.meta.env.VITE_BACKEND_URL
+}
+
+export default Configs
