+++ apps/server/.env.sample
@@ -1,3 +1,3 @@
 PORT = 4000
-WOOLF_GEMIN_ENDPOINT="<gemini proxy endpoint>"
+WOOLF_GEMIN_ENDPOINT="<woolf gemini proxy endpoint>"
 WOOLF_GEMINI_TOKEN="<authentication token>"
