+++ apps/server/package.json
@@ -15,6 +15,7 @@
     "@types/body-parser": "^1.19.6",
     "@types/cors": "^2.8.19",
     "@types/express": "^5.0.3",
+    "@types/morgan": "^1.9.10",
     "@types/multer": "^1.4.13",
     "@types/node": "^24.0.7",
     "@types/pdf-parse": "^1.1.5",
@@ -30,6 +31,8 @@
     "cors": "^2.8.5",
     "dotenv": "^17.0.0",
     "express": "^5.1.0",
+    "express-rate-limit": "^7.5.1",
+    "morgan": "^1.10.0",
     "multer": "^2.0.1",
     "pdf-parse": "^1.1.1",
     "pdf-parser": "^1.0.5",
