+++ apps/server/src/routes/rest/upload.route.ts
@@ -33,50 +33,4 @@ router.post("/",
     uploadController.uploadFile
 )
 
-router.post("/analyze",async(req,res)=>{
-
-    const data = req.body;
-    console.log("body: ",data)
-    const prompt = JSON.stringify({
-        data: {
-            cvText:data.cvText,
-            jdText:data.jdText
-        }
-    });
-
-
-    const systemInstruction = `You are an expert job recruiter
-
-You are tasked with analyzing job description and resume or cv of candidate given in the data object
-
-List down the strengths and weakness of the candidate and rate the candidate between 0 to 100 and give the final verdict if the candidate is a good match for the job or not.
-
-
-
-Give the output in JSON format
-Do not add extra characters or line breaks 
-example:{
-
-name: string //candidate name
-
-verdict: "Reject" | "Borderline | "Shortlist" | "Ideal",
-verdict_reason: "string"  // explain in 1 or 2 explanation for verdict
-
-matching_score: 45 (0-100)
-
-strengths: string[], // array of string containing strength points (maximum 6)
- 
-weakness: string[] //array of string containg weakness (maximum 6)
-}
-
-
-data: ${prompt}
-}`
-    const response =await  ai.Invoke(systemInstruction);
-    console.log(response);
-    return generateSuccessResponse(res,
-        JSON.parse(response)
-    );
-});
-
 export default router;
