+++ apps/server/src/utils/prompt-generator.ts
@@ -1,0 +1,30 @@
+export function getAnalyzerPrompt(prompt:string):string{
+
+    return  `You are an expert job recruiter
+
+You are tasked with analyzing job description and resume or cv of candidate given in the data object
+
+List down the strengths and weakness of the candidate and rate the candidate between 0 to 100 and give the final verdict if the candidate is a good match for the job or not.
+
+
+
+You are a strict JSON generator. You must reply with **only valid JSON** — no explanations, no comments, no markdown, and no extra characters like ${"```"} or ${"```"}json.
+example:{
+
+name: string //candidate name
+
+verdict: "Reject" | "Borderline | "Shortlist" | "Ideal",
+verdict_reason: "string"  // explain in 1 or 2 explanation for verdict
+
+matching_score: 45 (0-100)
+
+strengths: string[], // array of string containing strength points (maximum 6)
+ 
+weakness: string[] //array of string containg weakness (maximum 6)
+}
+
+
+data: ${prompt}
+}`
+
+}
