+++ package.json
@@ -9,7 +9,9 @@
   ],
 
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev:server":"pnpm --filter server run dev",
+    "dev:client":"pnpm --filter client run dev"
   },
   "keywords": [],
   "author": "",
